+++ deepctr/__init__.py
@@ -1,6 +1,4 @@
-from . import layers
-from . import models
 from .utils import check_version
 
-__version__ = '0.6.0'
+__version__ = '0.7.4'
 check_version(__version__)