+++ deepctr/layers/__init__.py
@@ -4,12 +4,13 @@ from .activation import Dice
 from .core import DNN, LocalActivationUnit, PredictionLayer
 from .interaction import (CIN, FM, AFMLayer, BiInteractionPooling, CrossNet,
                           InnerProductLayer, InteractingLayer,
-                          OutterProductLayer, FGCNNLayer,SENETLayer,BilinearInteraction)
+                          OutterProductLayer, FGCNNLayer,SENETLayer,BilinearInteraction,
+                          FieldWiseBiInteraction)
 from .normalization import LayerNormalization
 from .sequence import (AttentionSequencePoolingLayer, BiasEncoding, BiLSTM,
-                       KMaxPooling, SequencePoolingLayer,
+                       KMaxPooling, SequencePoolingLayer,WeightedSequenceLayer,
                        Transformer, DynamicGRU)
-from .utils import NoMask, Hash,Linear
+from .utils import NoMask, Hash,Linear,Add
 
 custom_objects = {'tf': tf,
                   'InnerProductLayer': InnerProductLayer,
@@ -38,4 +39,7 @@ custom_objects = {'tf': tf,
                   'DynamicGRU': DynamicGRU,
                   'SENETLayer':SENETLayer,
                   'BilinearInteraction':BilinearInteraction,
+                  'WeightedSequenceLayer':WeightedSequenceLayer,
+                  'Add':Add,
+                  'FieldWiseBiInteraction':FieldWiseBiInteraction
                   }