+++ deepctr/layers/sequence.py
@@ -14,12 +14,13 @@ from tensorflow.python.keras.layers import LSTM, Lambda, Layer
 
 from .core import LocalActivationUnit
 from .normalization import LayerNormalization
+
 if tf.__version__ >= '2.0.0':
     from ..contrib.rnn_v2 import dynamic_rnn
 else:
     from ..contrib.rnn import dynamic_rnn
 from ..contrib.utils import QAAttGRUCell, VecAttGRUCell
-from .utils import reduce_sum,reduce_max,div,softmax,reduce_mean
+from .utils import reduce_sum, reduce_max, div, softmax, reduce_mean
 
 
 class SequencePoolingLayer(Layer):
@@ -46,7 +47,7 @@ class SequencePoolingLayer(Layer):
         if mode not in ['sum', 'mean', 'max']:
             raise ValueError("mode must be sum or mean")
         self.mode = mode
-        self.eps = 1e-8
+        self.eps = tf.constant(1e-8, tf.float32)
         super(SequencePoolingLayer, self).__init__(**kwargs)
 
         self.supports_masking = supports_masking
@@ -63,7 +64,7 @@ class SequencePoolingLayer(Layer):
                 raise ValueError(
                     "When supports_masking=True,input must support masking")
             uiseq_embed_list = seq_value_len_list
-            mask = tf.cast(mask,tf.float32)#                tf.to_float(mask)
+            mask = tf.cast(mask, tf.float32)  # tf.to_float(mask)
             user_behavior_length = reduce_sum(mask, axis=-1, keep_dims=True)
             mask = tf.expand_dims(mask, axis=2)
         else:
@@ -77,15 +78,14 @@ class SequencePoolingLayer(Layer):
 
         mask = tf.tile(mask, [1, 1, embedding_size])
 
-        uiseq_embed_list *= mask
-        hist = uiseq_embed_list
         if self.mode == "max":
+            hist = uiseq_embed_list - (1-mask) * 1e9
             return reduce_max(hist, 1, keep_dims=True)
 
-        hist = reduce_sum(hist, 1, keep_dims=False)
+        hist = reduce_sum(uiseq_embed_list * mask, 1, keep_dims=False)
 
         if self.mode == "mean":
-            hist = div(hist, user_behavior_length + self.eps)
+            hist = div(hist, tf.cast(user_behavior_length, tf.float32) + self.eps)
 
         hist = tf.expand_dims(hist, axis=1)
         return hist
@@ -102,6 +102,83 @@ class SequencePoolingLayer(Layer):
     def get_config(self, ):
         config = {'mode': self.mode, 'supports_masking': self.supports_masking}
         base_config = super(SequencePoolingLayer, self).get_config()
+        return dict(list(base_config.items()) + list(config.items()))
+
+
+class WeightedSequenceLayer(Layer):
+    """The WeightedSequenceLayer is used to apply weight score on variable-length sequence feature/multi-value feature.
+
+      Input shape
+        - A list of two  tensor [seq_value,seq_len,seq_weight]
+
+        - seq_value is a 3D tensor with shape: ``(batch_size, T, embedding_size)``
+
+        - seq_len is a 2D tensor with shape : ``(batch_size, 1)``,indicate valid length of each sequence.
+
+        - seq_weight is a 3D tensor with shape: ``(batch_size, T, 1)``
+
+      Output shape
+        - 3D tensor with shape: ``(batch_size, T, embedding_size)``.
+
+      Arguments
+        - **weight_normalization**: bool.Whether normalize the weight score before applying to sequence.
+
+        - **supports_masking**:If True,the input need to support masking.
+    """
+
+    def __init__(self, weight_normalization=True, supports_masking=False, **kwargs):
+        super(WeightedSequenceLayer, self).__init__(**kwargs)
+        self.weight_normalization = weight_normalization
+        self.supports_masking = supports_masking
+
+    def build(self, input_shape):
+        if not self.supports_masking:
+            self.seq_len_max = int(input_shape[0][1])
+        super(WeightedSequenceLayer, self).build(
+            input_shape)  # Be sure to call this somewhere!
+
+    def call(self, input_list, mask=None, **kwargs):
+        if self.supports_masking:
+            if mask is None:
+                raise ValueError(
+                    "When supports_masking=True,input must support masking")
+            key_input, value_input = input_list
+            mask = tf.expand_dims(mask[0], axis=2)
+        else:
+            key_input, key_length_input, value_input = input_list
+            mask = tf.sequence_mask(key_length_input,
+                                    self.seq_len_max, dtype=tf.bool)
+            mask = tf.transpose(mask, (0, 2, 1))
+
+        embedding_size = key_input.shape[-1]
+
+        if self.weight_normalization:
+            paddings = tf.ones_like(value_input) * (-2 ** 32 + 1)
+        else:
+            paddings = tf.zeros_like(value_input)
+        value_input = tf.where(mask, value_input, paddings)
+
+        if self.weight_normalization:
+            value_input = softmax(value_input, dim=1)
+
+        if len(value_input.shape) == 2:
+            value_input = tf.expand_dims(value_input, axis=2)
+            value_input = tf.tile(value_input, [1, 1, embedding_size])
+
+        return tf.multiply(key_input, value_input)
+
+    def compute_output_shape(self, input_shape):
+        return input_shape[0]
+
+    def compute_mask(self, inputs, mask):
+        if self.supports_masking:
+            return mask[0]
+        else:
+            return None
+
+    def get_config(self, ):
+        config = {'weight_normalization': self.weight_normalization, 'supports_masking': self.supports_masking}
+        base_config = super(WeightedSequenceLayer, self).get_config()
         return dict(list(base_config.items()) + list(config.items()))
 
 
@@ -383,11 +460,11 @@ class Transformer(Layer):
         self.supports_masking = supports_masking
 
     def build(self, input_shape):
-
         embedding_size = int(input_shape[0][-1])
         if self.num_units != embedding_size:
             raise ValueError(
-                "att_embedding_size * head_num must equal the last dimension size of inputs,got %d * %d != %d" % (self.att_embedding_size,self.head_num,embedding_size))
+                "att_embedding_size * head_num must equal the last dimension size of inputs,got %d * %d != %d" % (
+                self.att_embedding_size, self.head_num, embedding_size))
         self.seq_len_max = int(input_shape[0][-2])
         self.W_Query = self.add_weight(name='query', shape=[embedding_size, self.att_embedding_size * self.head_num],
                                        dtype=tf.float32,
@@ -467,10 +544,10 @@ class Transformer(Layer):
         if self.blinding:
             try:
                 outputs = tf.matrix_set_diag(outputs, tf.ones_like(outputs)[
-                                                  :, :, 0] * (-2 ** 32 + 1))
+                                                      :, :, 0] * (-2 ** 32 + 1))
             except:
                 outputs = tf.compat.v1.matrix_set_diag(outputs, tf.ones_like(outputs)[
-                                                      :, :, 0] * (-2 ** 32 + 1))
+                                                                :, :, 0] * (-2 ** 32 + 1))
 
         outputs -= reduce_max(outputs, axis=-1, keep_dims=True)
         outputs = softmax(outputs)
@@ -519,6 +596,7 @@ class Transformer(Layer):
                   'blinding': self.blinding}
         base_config = super(Transformer, self).get_config()
         return dict(list(base_config.items()) + list(config.items()))
+
 
 def positional_encoding(inputs,
                         pos_embedding_trainable=True,
@@ -708,7 +786,7 @@ class KMaxPooling(Layer):
 
         if self.axis < 1 or self.axis > len(input_shape):
             raise ValueError("axis must be 1~%d,now is %d" %
-                             (len(input_shape), len(input_shape)))
+                             (len(input_shape), self.axis))
 
         if self.k < 1 or self.k > input_shape[self.axis]:
             raise ValueError("k must be in 1 ~ %d,now k is %d" %