+++ deepctr/models/onn.py
@@ -14,7 +14,7 @@ import itertools
 import tensorflow as tf
 from tensorflow.python.keras import backend as K
 from tensorflow.python.keras.initializers import RandomNormal
-from tensorflow.python.keras.layers import (Dense, Embedding, Lambda, add,
+from tensorflow.python.keras.layers import (Dense, Embedding, Lambda,
                                             multiply)
 from tensorflow.python.keras.models import Model
 from tensorflow.python.keras.regularizers import l2
@@ -23,13 +23,13 @@ from ..inputs import (build_input_features, VarLenSparseFeat,
                       get_linear_logit, SparseFeat, get_dense_input, combined_dnn_input)
 from ..layers.core import DNN, PredictionLayer
 from ..layers.sequence import SequencePoolingLayer
-from ..layers.utils import concat_fun, Hash, NoMask
+from ..layers.utils import concat_func, Hash, NoMask, add_func
 
 
-def NFFM(linear_feature_columns, dnn_feature_columns, embedding_size=4, dnn_hidden_units=(128, 128),
-         l2_reg_embedding=1e-5, l2_reg_linear=1e-5, l2_reg_dnn=0, dnn_dropout=0,
-         init_std=0.0001, seed=1024, use_bn=True, reduce_sum=False, task='binary',
-         ):
+def ONN(linear_feature_columns, dnn_feature_columns, embedding_size=4, dnn_hidden_units=(128, 128),
+        l2_reg_embedding=1e-5, l2_reg_linear=1e-5, l2_reg_dnn=0, dnn_dropout=0,
+        init_std=0.0001, seed=1024, use_bn=True, reduce_sum=False, task='binary',
+        ):
     """Instantiates the Operation-aware Neural Networks  architecture.
 
     :param linear_feature_columns: An iterable containing all the features used by linear part of the model.
@@ -53,15 +53,15 @@ def NFFM(linear_feature_columns, dnn_feature_columns, embedding_size=4, dnn_hidd
 
     inputs_list = list(features.values())
 
-    linear_logit = get_linear_logit(features, linear_feature_columns, l2_reg=l2_reg_linear, init_std=init_std,
-                                    seed=seed, prefix='linear')
+    linear_logit = get_linear_logit(features, linear_feature_columns, init_std=init_std, seed=seed, prefix='linear',
+                                    l2_reg=l2_reg_linear)
 
     sparse_feature_columns = list(
         filter(lambda x: isinstance(x, SparseFeat), dnn_feature_columns)) if dnn_feature_columns else []
     varlen_sparse_feature_columns = list(
         filter(lambda x: isinstance(x, VarLenSparseFeat), dnn_feature_columns)) if dnn_feature_columns else []
 
-    sparse_embedding = {fc_j.embedding_name: {fc_i.embedding_name: Embedding(fc_j.dimension, embedding_size,
+    sparse_embedding = {fc_j.embedding_name: {fc_i.embedding_name: Embedding(fc_j.vocabulary_size, embedding_size,
                                                                              embeddings_initializer=RandomNormal(
                                                                                  mean=0.0, stddev=0.0001, seed=seed),
                                                                              embeddings_regularizer=l2(
@@ -80,10 +80,10 @@ def NFFM(linear_feature_columns, dnn_feature_columns, embedding_size=4, dnn_hidd
     for fc_i, fc_j in itertools.combinations(sparse_feature_columns + varlen_sparse_feature_columns, 2):
         i_input = features[fc_i.name]
         if fc_i.use_hash:
-            i_input = Hash(fc_i.dimension)(i_input)
+            i_input = Hash(fc_i.vocabulary_size)(i_input)
         j_input = features[fc_j.name]
         if fc_j.use_hash:
-            j_input = Hash(fc_j.dimension)(j_input)
+            j_input = Hash(fc_j.vocabulary_size)(j_input)
 
         fc_i_embedding = feature_embedding(fc_i, fc_j, sparse_embedding, i_input)
         fc_j_embedding = feature_embedding(fc_j, fc_i, sparse_embedding, j_input)
@@ -94,21 +94,15 @@ def NFFM(linear_feature_columns, dnn_feature_columns, embedding_size=4, dnn_hidd
                 element_wise_prod, axis=-1))(element_wise_prod)
         embed_list.append(element_wise_prod)
 
-    ffm_out = tf.keras.layers.Flatten()(concat_fun(embed_list, axis=1))
+    ffm_out = tf.keras.layers.Flatten()(concat_func(embed_list, axis=1))
     if use_bn:
         ffm_out = tf.keras.layers.BatchNormalization()(ffm_out)
     dnn_input = combined_dnn_input([ffm_out], dense_value_list)
     dnn_out = DNN(dnn_hidden_units, l2_reg=l2_reg_dnn, dropout_rate=dnn_dropout)(dnn_input)
     dnn_logit = Dense(1, use_bias=False)(dnn_out)
 
-    if len(linear_feature_columns) > 0 and len(dnn_feature_columns) > 0:
-        final_logit = add([dnn_logit, linear_logit])
-    elif len(linear_feature_columns) > 0:
-        final_logit = linear_logit
-    elif len(dnn_feature_columns) > 0:
-        final_logit = dnn_logit
-    else:
-        raise NotImplementedError
+
+    final_logit = add_func([dnn_logit,linear_logit])
 
     output = PredictionLayer(task)(final_logit)